Add files via upload
Basically run the read_votes program with all other Python files in one common folder. You will only see one picture with weird bars on it. Boring, isn't it? But it was worth 7% of my final mark. I guess it is simply a deep practice of lists, dictionary, functions, etc.